[Fuzzer] Move all network related fuzzing targets into same file.
