+++ testsuite/libra-fuzzer/src/fuzz_targets.rs
@@ -14,9 +14,7 @@ mod inner_signed_transaction;
 mod json_rpc_service;
 mod language_transaction_execution;
 mod mempool;
-mod network_noise_initiator;
-mod network_noise_responder;
-mod network_noise_stream;
+mod network_noise;
 mod state_sync_msg;
 //mod storage_save_blocks;
 mod execute_and_commit_chunk;
@@ -32,9 +30,9 @@ static ALL_TARGETS: Lazy<BTreeMap<&'static str, Box<dyn FuzzTargetImpl>>> = Lazy
         Box::new(inner_signed_transaction::SignedTransactionTarget::default()),
         Box::new(json_rpc_service::JsonRpcSubmitTransactionRequest::default()),
         Box::new(mempool::MempoolIncomingTransactions::default()),
-        Box::new(network_noise_initiator::NetworkNoiseInitiator::default()),
-        Box::new(network_noise_responder::NetworkNoiseResponder::default()),
-        Box::new(network_noise_stream::NetworkNoiseStream::default()),
+        Box::new(network_noise::NetworkNoiseInitiator::default()),
+        Box::new(network_noise::NetworkNoiseResponder::default()),
+        Box::new(network_noise::NetworkNoiseStream::default()),
         Box::new(state_sync_msg::StateSyncMsg::default()),
         Box::new(language_transaction_execution::LanguageTransactionExecution::default()),
         //        Box::new(storage_save_blocks::StorageSaveBlocks::default()),
